Introduce ProxyCredentials to better enforce the nullability contract
