+++ gradle/javaModule.gradle
@@ -68,7 +68,7 @@ tasks.withType(Test).configureEach {
   systemProperty "io.netty.allocator.type", "unpooled"
   systemProperty "io.netty.leakDetectionLevel", "paranoid"
 
-  jvmArgs "-Xss320k", "-XX:MaxDirectMemorySize=64m"
+  jvmArgs "-Xss320k", "-XX:MaxDirectMemorySize=128m"
   minHeapSize "256m"
   maxHeapSize "768m"
