+++ ratpack-core/src/main/java/ratpack/http/client/ProxyCredentials.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package ratpack.http.client;
+
+/**
+ * Credentials to use when using a HTTP proxy that requires authentication.
+ */
+public interface ProxyCredentials {
+
+  /**
+   * The username to use when connecting to the proxy.
+   *
+   * @return The username to use when connecting to the HTTP proxy.
+   */
+  String getUsername();
+
+  /**
+   * The password to use when connecting to the proxy.
+   *
+   * @return Either the password to use when connecting to the HTTP proxy.
+   */
+  String getPassword();
+
+}
