+++ ratpack-core/src/main/java/ratpack/http/client/ProxySpec.java
@@ -16,8 +16,6 @@
 
 package ratpack.http.client;
 
-import ratpack.api.Nullable;
-
 import java.util.Collection;
 
 /**
@@ -42,20 +40,13 @@ public interface ProxySpec {
   ProxySpec port(int port);
 
   /**
-   * Configure the username to use when connecting to the proxy.
+   * Configure the username and password to use when connecting to the proxy.
    *
    * @param username the username to use when connecting to the HTTP proxy
-   * @return {@code this}
-   */
-  ProxySpec username(@Nullable String username);
-
-  /**
-   * Configure the password to use when connecting to the proxy.
-   *
    * @param password the password to use when connecting to the HTTP proxy
    * @return {@code this}
    */
-  ProxySpec password(@Nullable String password);
+  ProxySpec credentials(String username, String password);
 
   /**
    * Configure a collection of patterns for which if any are matched, the outbound request will bypass the HTTP proxy.
