+++ ratpack-core/src/main/java/ratpack/http/client/internal/DefaultProxyCredentials.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package ratpack.http.client.internal;
+
+import ratpack.http.client.ProxyCredentials;
+
+import static com.google.common.base.Preconditions.checkNotNull;
+
+public class DefaultProxyCredentials implements ProxyCredentials {
+
+  private final String username;
+  private final String password;
+
+  public DefaultProxyCredentials(String username, String password) {
+    checkNotNull(username);
+    checkNotNull(password);
+    this.username = username;
+    this.password = password;
+  }
+
+  @Override
+  public String getUsername() {
+    return username;
+  }
+
+  @Override
+  public String getPassword() {
+    return password;
+  }
+}
